Add returning price for one night and total price to the filter

+++ src/main/java/com/booking/project/dto/AccommodationCardDTO.java
@@ -18,8 +18,8 @@ public class AccommodationCardDTO {
     private Address address;
     private String description;
     private Double avgRate;
-    private Integer pricePerNight;
-    private Integer totalPrice;
+    private Double pricePerNight;
+    private Double totalPrice;
     private byte[] image;
 
     public AccommodationCardDTO(Accommodation accommodation){
